+++ Point.cpp
@@ -1,7 +1,7 @@
 #include "Point.hpp"
 #include <cmath>
 
-const double Point::EPS = 10e-6;
+const float Point::EPS = 10e-6;
 
 bool Point::operator==(const Point& p2) const {
 	return fabs(x - p2.x) < EPS && fabs(y - p2.y) < EPS && fabs(z - p2.z) < EPS;
@@ -14,7 +14,7 @@ std::ostream & operator<<(std::ostream& os, const Point& V1){
     return os << V1.x << ' ' << V1.y << ' ' << V1.z;
 }
 
-double Point::distance(const Point& p1, const Point& p2) {
+float Point::distance(const Point& p1, const Point& p2) {
 	return sqrt((p1.x - p2.x) * (p1.x - p2.x) +
 		(p1.y - p2.y) * (p1.y - p2.y) +
 		(p1.z - p2.z) * (p1.z - p2.z));
@@ -26,7 +26,7 @@ std::size_t PointHasher::operator()(const Point& p) const
 	using std::hash;
 	using std::string;
 
-	return ((hash<double>()(p.x)
-		^ (hash<double>()(p.y) << 1)) >> 1)
-		^ (hash<double>()(p.z) << 1);
+	return ((hash<float>()(p.x)
+		^ (hash<float>()(p.y) << 1)) >> 1)
+		^ (hash<float>()(p.z) << 1);
 }
